Debug Ublox driver
Should work on 32 bits system but bugued on 64 bits one (Bad value and big jumps).
Use sized type to be a portable code.

+++ sunray/src/ublox/ublox.h
@@ -1,11 +1,11 @@
 /*
-// Ardumower Sunray 
+// Ardumower Sunray
 // Copyright (c) 2013-2020 by Alexander Grau, Grau GmbH
 // Licensed GPLv3 for open source use
 // or Grau GmbH Commercial License for commercial use (http://grauonline.de/cms2/?page_id=153)
 
   ublox f9p UBX parser
-  uses UBX messages: 
+  uses UBX messages:
   1) 'UBX-NAV-PTV'        10  (Navigation Position Velocity Time Solution)
   2) 'UBX-NAV-RELPOSNED'   1  (Relative Positioning Information in NED frame)
   3) 'UBX-NAV-HPPOSLLH'    1  (absolute position)
@@ -19,12 +19,12 @@
 #ifndef UBLOX_h
 #define UBLOX_h
 
-#include "Arduino.h"				
+#include "Arduino.h"
 #include "../../gps.h"
 #include "../driver/RobotDriver.h"
 
 class UBLOX : public GpsDriver {
-  public:    
+  public:
     typedef enum {
         GOT_NONE,
         GOT_SYNC1,
@@ -32,21 +32,21 @@ class UBLOX : public GpsDriver {
         GOT_CLASS,
         GOT_ID,
         GOT_LENGTH1,
-        GOT_LENGTH2, 
+        GOT_LENGTH2,
         GOT_PAYLOAD,
-        GOT_CHKA 
+        GOT_CHKA
 
-    } state_t;        
+    } state_t;
     UBLOX();
     void begin(Client &client, char *host, uint16_t port) override;
     void begin(HardwareSerial& bus,uint32_t baud) override;
     void run() override;
-    bool configure() override;  
+    bool configure() override;
     void reboot() override;
   private:
     bool useTCP;
-    Client* _client;    
-    uint32_t _baud;  	
+    Client* _client;
+    uint32_t _baud;
     HardwareSerial* _bus;
     state_t state;
     int msgid;
@@ -55,19 +55,19 @@ class UBLOX : public GpsDriver {
     int chka;
     int chkb;
     int count;
-    char payload[2000];                                          
+    char payload[2000];
     bool debug;
     bool verbose;
-    unsigned long solutionTimeout;    
+    unsigned long solutionTimeout;
 
     void begin();
     void addchk(int b);
     void dispatchMessage();
-    long unpack_int32(int offset);
-    long unpack_int16(int offset);
-    long unpack_int8(int offset);
-    long unpack(int offset, int size);
-    void parse(int b);	  	    
+    int32_t unpack_int32(int offset);
+    int16_t unpack_int16(int offset);
+    int8_t unpack_int8(int offset);
+    int32_t unpack(int offset, int size);
+    void parse(int b);
 };
 
 #endif
